Implement item data helpers for normalization and querying
